@@ -12,17 +12,16 @@
   "license": "ISC",
   "devDependencies": {
     "@types/express": "^4.16.1",
-    "tslint": "^5.12.1",
-    "typescript": "^3.3.3"
+    "@types/redis": "^2.8.28",
+    "typescript": "^4.1.3",
+    "nodemon": "^2.0.7"
   },
   "dependencies": {
-    "@types/redis": "^2.8.28",
     "axios": "^0.21.1",
-    "bot-ts": "^0.2.1",
     "discord.js": "^12.5.1",
     "dotenv": "^8.2.0",
+    "eslint": "^7.18.0",
     "express": "^4.16.4",
-    "nodemon": "^2.0.7",
     "redis": "^3.0.2"
   }
 }
